Basic beach sprite
Adding a beach sprite which is responsible for drawing both the sand
and the waves, this is because the sand will want to know where the
water is in order to be damp, and the water will want to know about
the slope of the sand in order to figure out which x position counts
as a collision for the player.

Currently using a single tween to animate the waves, but this should
probably be done with delays and custom tweens for in and out instead
of relying on the yoyo.