Add wet sand under waves
Ok, so this is not a simple implementation, there's a lot of vector
maths to figure out where the wet sand should be drawn and to get it
to follow the wave in an intuitive way. The advantage is that if we
want to change the angle of the beach or the height of the sea
everything should just work out.

This would all probably be solved if I had been smart enough to just
do the whole thing horizontally and then rotate the whole game with a
single transform after. Whatever, too late to change now.